+++ Source/Widgets/Plotting/ORTimeSeriesPlot.m
@@ -99,14 +99,14 @@
 		x	= [mXScale getPixAbs:minChan]-chanWidth/2.;
 		xl	= x;
 		double xValue,yValue;
-		[dataSource plotter:self index:minChan x:&xValue y:&yValue];
+		[dataSource plotter:self index:(int)minChan x:&xValue y:&yValue];
 		yl	= [mYScale getPixAbs:yValue];
 		long ix;
 		for (ix=minChan; ix<=maxChan+1;++ix) {
 			if(ix > numPoints)break;
 			double xValue;
 			double yValue;
-			[dataSource plotter:self index:ix x:&xValue y:&yValue];
+			[dataSource plotter:self index:(int)ix x:&xValue y:&yValue];
 			
 			x = [mXScale getPixAbsFast:ix log:NO integer:YES minPad:aMinPadx] - chanWidth/2.;
 			y = [mYScale getPixAbsFast:yValue log:aLog integer:aInt minPad:aMinPad];
